// sp/ClientApp/src/app/models/messages/search-message-info.model.ts
export class SearchMessageInfoModel {
    Attachement: string;
    body: string;
    contactName: string;
    contactImage: string;
    senderImage: string;
    contactID: string;
    flagLevel: string;
    importanceLevel: string;
    messageID: string;
    messageState: string;
    senderID: string;
    subject: string;
    msgDate: string;
    fromID: string;
    firstName: string;
    fullBody: string;
    selected: boolean = false;

}